move severity test into reverse mapper test file, remove a duplicate test, and refactor test file
